small fixes ant use bundle.entry instead of bundle.items
