+++ client/src/components/mapping/TableViewer/style.scss
