+++ client/src/components/mapping/style.scss
@@ -49,6 +49,10 @@ body {
   }
 }
 
+#tableViewer {
+  padding-top: 5%;
+  padding-left: 20px;
+}
 .bp3-tab-list {
   background-color: $light-gray1;
   overflow: hidden;
