+++ client/src/components/selects/addResourceSelect.tsx
@@ -160,7 +160,7 @@ const AddResourceSelect = ({
         filterItems={filterByName}
         loading={loading}
         icon={icon}
-        inputItem={inputItem || ({ name: 'Select a resource' } as Resource)}
+        inputItem={inputItem || ({ name: 'Create a resource' } as Resource)}
         intent={intent}
         items={sortItems(items || []) as Resource[]}
         onChange={onChange}
